another type update to get by clang on mac

+++ src/audio.c
@@ -162,7 +162,7 @@ struct aurx {
 	size_t last_sampc;
 
 	struct {
-		uint32_t aubuf_current_jb;
+		uint64_t aubuf_current_jb;
 		uint64_t aubuf_overrun;
 		uint64_t aubuf_underrun;
 		uint64_t n_discard;
@@ -192,7 +192,7 @@ struct audio {
 /* RFC 6464 */
 static const char *uri_aulevel = "urn:ietf:params:rtp-hdrext:ssrc-audio-level";
 
-uint32_t audio_jb_current_value(const struct audio* au)
+uint64_t audio_jb_current_value(const struct audio* au)
 {
 	if (!au)
 		return 0;
@@ -618,7 +618,7 @@ static void auplay_write_handler(void *sampv, size_t sampc, void *arg)
 {
 	struct aurx *rx = arg;
 	size_t num_bytes = sampc * aufmt_sample_size(rx->play_fmt);
-	uint32_t b_p_ms = 0; /* bytes per ms */
+	uint64_t b_p_ms = 0; /* bytes per ms */
 
 	if (rx->aubuf_started && aubuf_cur_size(rx->aubuf) < num_bytes) {
 
@@ -630,9 +630,9 @@ static void auplay_write_handler(void *sampv, size_t sampc, void *arg)
 #endif
 	}
 	b_p_ms = aufmt_sample_size(rx->play_fmt)*
-		rx->auplay_prm.srate *rx->auplay_prm.ch / 1000;
+                rx->auplay_prm.srate *rx->auplay_prm.ch / 1000;
 	if (rx->stats.aubuf_current_jb != aubuf_cur_size(rx->aubuf)/b_p_ms) {
-		rx->stats.aubuf_current_jb = aubuf_cur_size(rx->aubuf)/b_p_ms;
+                rx->stats.aubuf_current_jb = aubuf_cur_size(rx->aubuf)/b_p_ms;
 	}
 	aubuf_read(rx->aubuf, sampv, num_bytes);
 }
